refactor: Controller 분리로 if 분기 제거
